Refactor response objects into new package
