+++ app/app_test.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"github.com/mortedecai/gbb/gbb"
 	"github.com/mortedecai/gbb/gbb/mocks"
+	response2 "github.com/mortedecai/gbb/gbb/response"
 	"go.uber.org/mock/gomock"
 	"io"
 	"net/http"
@@ -122,10 +123,10 @@ var _ = Describe("App", func() {
 })
 
 func createEmptyDownloadResponse() *http.Response {
-	basicResponse := gbb.GBBDownloadFilesResponse{
+	basicResponse := response2.GBBDownloadFilesResponse{
 		Success: true,
-		Data: gbb.GBBFileDownloadData{
-			Files: make([]gbb.GBBDownloadFile, 0),
+		Data: response2.GBBFileDownloadData{
+			Files: make([]response2.GBBDownloadFile, 0),
 		},
 	}
 	data, err := json.Marshal(basicResponse)
